+++ NEWS.md
@@ -1,3 +1,94 @@
+# modelbpp 0.1.5
+
+## New Features
+
+- Added the argument `exclude_xy_cov`
+  and `exclude_feedback` to `get_add()`
+  and `model_set()`, for excluding paths
+  that create feedback loops, and
+  covariances involving a predictor
+  and an outcome variable (including
+  those linked by indirect paths).
+  Default values has been changed
+  to `TRUE` since 0.1.3.5. To
+  reproduce results from previous version,
+  set them to `FALSE`. (0.1.3.2, 0.1.3.5)
+
+- Added `min_bpp_labelled` to
+  `model_graph()`, to hide the labels
+  of models with small BPPs.
+  (0.1.3.5)
+
+- Added the argument `drop_equivalent_models`,
+  to `model_set()`. If `TRUE`, the
+  default, the models fitted will be
+  checked for equivalence. If two or
+  more more models are equivalent, only
+  one of them will be retained.
+  The groups of equivalent models identified,
+  and the models dropped, will be
+  printed by the print method. (0.1.3.9)
+
+- Added `measurement_invariance_models()`,
+  for generating metric and scalar
+  invariance models and their partial
+  invariance versions. (0.1.3.10 - 0.1.3.11)
+
+## Miscellaneous
+
+- Because it is very likely that users
+  would like to see come fit measures
+  along with BPPs, the default of
+  `more_fit_measures` of the print
+  method of `model_set`-class object
+  changed to `c("cfi", "rmsea")`.
+  (0.1.3.7)
+
+- Revised `fit_many()` to support
+  multigroup models. (0.1.3.8)
+
+- A progress bar can be displayed when
+  `model_set()` is identifying nested
+  models. (0.1.3.13)
+
+- Shortened `BIC Posterior Probability`
+  to `BPP` in some sections of the
+  printout of `print.model_set()`.
+  (0.1.3.14)
+
+- Cumulative BPPs no longer displayed
+  by default in `print.model_set()`.
+  Print them by setting `cumulative_bpp`
+  to `TRUE`. (0.1.3.15)
+
+- Update an internal function to handle
+  nonconvergence in checking nested
+  relation. Only affect the graphs and
+  only happen in some rare cases.
+  (0.1.3.16)
+
+## Bug Fixes
+
+- The `must_not_add` argument should
+  work now for some parameters that may
+  not be recognized as interchangeable.
+  (0.1.3.1)
+
+- Fixed a bug in `must_not_drop` and
+  `must_drop` of `get_drop()`. They
+  should work properly now. (0.1.3.5)
+
+- Fixed a bug in `model_graph()`.
+  Short names should now be properly
+  constructed. (0.1.3.3)
+
+- Fixed some bugs in `print.model_set()`
+  about the printing of additional fit
+  measures. (0.1.3.6, 0.1.3.7)
+
+- Fixed a bug in checking whether two
+  models are equivalent. (0.1.3.12)
+
 # modelbpp 0.1.3
 
 ## New Features